Hardcode the zm namelist parameters, additional cleanup of cam dependencies.
This commit replaces the zm namelist reading with hard-coded values.

It also circumvents subroutines in ZM that may not be needed.

Finally, it replaces the is_first_step usage from the time manager with
a local value.  This will need to be changed back once teh AD is able to
provide this information.